requirements.txt
asgiref==3.3.4
astroid==2.5.6
Django==3.2.1
django-crispy-forms==1.11.2
django-extensions==3.1.3
django-phone-field==1.8.1
django-phonenumber-field==5.1.0
django-quill-editor==0.1.22
django-taggit==1.4.0
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
phonenumbers==8.12.23
Pillow==8.2.0
pydot==1.4.2
pygraphviz==1.7
pylint==2.8.2
pyparsing==2.4.7
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
wrapt==1.12.1
